refactor: Board, Game 분리
- Game은 체스판의 점수를 관리한다
